Moved adding to file into a spearate function

+++ web_server_with_database/src/common_utils/global_types.rs
@@ -7,7 +7,7 @@ pub struct SubmittedOrderData {
     pub email: String,
     pub copies_nbr: u32,
     pub file_name: String,
-    pub nbr_of_chunks: usize
+    pub nbr_of_chunks: u32
 }
 
 pub struct EvaluationResult {
